Move change detection code to separate class
